@@ -5,7 +5,7 @@ author: Alban Veliu <alveliu93<@gmail.com>
 homepage: https://github.com/albo1337/flutter_full_pdf_viewer
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
